chore: construct project stub
This commit creates the project stub
using methods as minimalistic as possible:
The npm utility was used to generate the package file.
The tsc utility was used to generate the tsconfig file.
Prettier and linter were configured via plugins.
The eslint file extends the plugins
typescript-eslint/recommended and prettier/recommended.
The tsconfig file was configured to describe a library,
that is, target was set to es5 (for isomorphism),
the out dir was set to the lib directory,
and the module was set to commonjs.
Jest was installed along with ts-jest
to support typescript and es6 imports.
Simple gitignore file was created to exclude
compiled output, os files and IDE files
(excluding settings and extension recommendations
for this project).
The README file was put into the docs directory,
as well as the CONTRIBUTING file,
which is linked from the latter.
Example files and tests were added,
so it is possible to
build, run unit tests or lint.